% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animatedTimeSeries.R
\name{animatedTimeSeries}
\alias{animatedTimeSeries}
\title{<Add Title>}
\usage{
animatedTimeSeries(
  data,
  settings = NULL,
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\description{
<Add Description>
}
